refactor(collab-plane): share WsCloseCode, extract broadcastJson, use event constants
